requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
anthropic>=0.28.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pypdf>=4.0.0
python-docx>=1.1.0
